// src/models/CasePrimary.ts
export type CasePrimaryResponse = {
    id: number;
    caseRefNum: string;
    codeCaseStatus: string;
    lodgeDate: Date;
    incidentDate: Date;
    iniRptDate: string;
    codeRptMtd: string;
    codeCapoFmt: string;
    codeInvTeam: string;
    codeEnqUnit: string;
    codeComOption: string;
    indParaInvByOther: string;
    indImRel: string;
    indJrSuitRel: string;
    indPressInt: string;
    indViceRel: string;
    indSentitivePerson: string;
    indSensitiveLoc: string;
    indMipRel: string;
    indSpHandling: string;
    categarization: string;
    complaintType: string;
}

export const emptyCasePrimaryResponse: CasePrimaryResponse = {
    id: 0,                                      
    caseRefNum: "",
    codeCaseStatus: "",
    lodgeDate: null,
    incidentDate: null,
    iniRptDate: "",
    codeRptMtd: "",
    codeCapoFmt: "",
    codeInvTeam: "",
    codeEnqUnit: "",
    codeComOption: "",
    indParaInvByOther: "",
    indImRel: "",
    indJrSuitRel: "",
    indPressInt: "",
    indViceRel: "",
    indSentitivePerson: "",
    indSensitiveLoc: "",
    indMipRel: "",
    indSpHandling: "",
    categarization: "",
    complaintType: "",
}